requirements.txt
yfinance==0.2.36
fredapi==0.5.0
pandas==2.2.0
numpy==1.26.3
scipy==1.11.4
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
dbt-core==1.7.9
dbt-postgres==1.7.9
jupyterlab==4.1.0
python-dotenv==1.0.1
jupyterlab
matplotlib
seaborn
psycopg2-binary
nbformat
